Changed inventoryscript & health script
made it so when you use seeds it heals player, also now the 'pink' seed gives you a 'superseed' which just heals for double a normal seed (right now they health for 20 health each and the super seed 40). Also added sound to trigger when you eat a seed.
I added a function to the health script called "AddHealth" you can change it salick if you want, just did it for now :+1: